Expense item price converted to double
- Was having trouble when converting string to float with different number format locale.

+++ Troshko/Modules/MonthlyOverview/Models/MonthlyOverviewModel.swift
@@ -9,5 +9,5 @@ import Foundation
 
 struct MonthlyOverviewModel {
     var categoryName: String
-    var totalExpenses: Float
+    var totalExpenses: Double
 }
